parser/requirements.txt
beautifulsoup4
urllib3
Psycopg2-binary
sockets
git